base: Add tests for abstract domains
Fixes #617.
